% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_comparison_table.R
\name{model_comparison_table}
\alias{model_comparison_table}
\title{model_comparison_table
Takes in training data, testing data,
with the target as the last column and
fitted models with meaningful names,
then generates a model comparison
table.}
\usage{
model_comparison_table(train_data, test_data, ...)
}
\arguments{
\item{train_data}{tibble of training data with target as last column}

\item{test_data}{tibble of testing data with target as last column}

\item{...}{fitted models assigned to meaningful model names}
}
\value{
tibble of results, allowing you to compare models
}
\description{
model_comparison_table
Takes in training data, testing data,
with the target as the last column and
fitted models with meaningful names,
then generates a model comparison
table.
}
\examples{
\dontrun{
library(RMLViz)
library(mlbench)
data(Sonar)

toy_classification_data <- dplyr::select(dplyr::as_tibble(Sonar), V1, V2, V3, V4, V5, Class)

train_ind <- caret::createDataPartition(toy_classification_data$Class, p=0.9, list=F)
train_set_cf <- toy_classification_data[train_ind, ]
test_set_cf <- toy_classification_data[-train_ind, ]

gbm <- caret::train(Class~., train_set_cf, method="gbm", verbose=F)
lm_cf <- caret::train(Class~., train_set_cf, method="LogitBoost", verbose=F)

model_comparison_table(train_set_cf, test_set_cf,
                       gbm_mod=gbm, log_mod = lm_cf)
}
}
