% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roc.R
\name{plot_roc}
\alias{plot_roc}
\title{creates a ROC plot}
\usage{
plot_roc(y_label, predict_proba)
}
\arguments{
\item{y_label}{a numeric vector contains the validation set label, all elements should be either 0 or 1}

\item{predict_proba}{a numberic vector contains the prediction probabilities}
}
\value{
A plot
}
\description{
Takes in a vector of prediction labels and a vector of prediction probabilities.
}
\examples{
\dontrun{
library(RMLViz)
set.seed(420)

num.samples <- 100
weight <- sort(rnorm(n=num.samples, mean=172, sd=29))
bese <- ifelse(test=(runif(n=num.samples) < (rank(weight)/num.samples)),
               yes=1, no=0)

glm.fit=glm(obese ~ weight, family=binomial)
obese_proba <- glm.fit$fitted.values

plot_roc(obese, obese_proba)
}
}
