% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_train_valid_error.R
\name{plot_train_valid_error}
\alias{plot_train_valid_error}
\title{plot train/validation errors vs. parameter values}
\usage{
plot_train_valid_error(
  model_name,
  X_train,
  y_train,
  X_valid,
  y_valid,
  param_name,
  param_vec
)
}
\arguments{
\item{model_name}{a string of the machine learning model name.
Only 'knn', 'decision tree', 'svm', and 'random forests' are allowed.}

\item{X_train}{a numeric data frame of the training dataset without labels.}

\item{y_train}{a numeric vector or factor of the training labels.}

\item{X_valid}{a numeric data frame of the validation dataset without labels.}

\item{y_valid}{a numeric vector or factor of the validation labels.}

\item{param_name}{a string of the parameter name.
Please choose this parameter based on the following information:
'knn': 'k',
'decision tree': 'maxdepth',
'svm': 'cost' or 'gamma',
'random forests': 'ntree'.}

\item{param_vec}{a numeric vector of the parameter values.}
}
\value{
A plot
}
\description{
Takes in a model name, train/validation data sets,
a parameter name and a vector of parameter values
to try and then plots train/validation errors vs.
parameter values.
}
\examples{
\dontrun{
library(RMLViz)
data(iris)
set.seed(123)

split <- caTools::sample.split(iris$Species, SplitRatio = 0.75)
training_set <- subset(iris, split == TRUE)
valid_set <- subset(iris, split == FALSE)

X_train <- training_set[, -5]
y_train <- training_set[, 5]
X_valid <- valid_set[, -5]
y_valid <- valid_set[, 5]

plot_train_valid_error('knn',
                       X_train, y_train,
                       X_valid, y_valid,
                       'k', seq(50))
}
}
